+++ Homework/opt/midterm/ex2/linkedlist.h
@@ -3,12 +3,13 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <stdint.h>
+#include <stdbool.h>
 
 typedef struct node_t{
     struct node_t *next;
-    int start;
-    int end;
+    double start;
+    double end;
 } node_t;
 
-node_t *push(node_t *head, int start, int end);
+node_t *push(node_t *head, double start, double end);
 node_t *destroy_list(node_t *head);
